ingest/requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
dateparser==1.1.8
python-dateutil==2.8.2
beautifulsoup4==4.12.2
html2text==2020.1.16
requests==2.31.0
